pyproject.toml
[project]
name = "panekmodel2"
version = "0.1.0"
description = "YouTube transcript ingestion, topic modeling (BERTopic), and sentiment pipeline"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9.0",
  "rich>=13.7.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.6.0",
  "youtube-transcript-api>=0.6.2",
  "google-api-python-client>=2.154.0",
  "google-auth-oauthlib>=1.2.0",
  "pytube>=15.0.0",
  "openai-whisper>=20240930",
  "torch>=2.1.0",
  "pandas>=2.1.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.3.0",
  "bertopic>=0.16.0",
  "hdbscan>=0.8.33",
  "umap-learn>=0.5.5",
  "sentence-transformers>=2.7.0",
  "transformers>=4.39.0",
  "tqdm>=4.66.0",
  "nltk>=3.8.1",
  "matplotlib>=3.8.0",
  "srt>=3.5.0",
  "yt-dlp>=2024.11.4",
  "streamlit>=1.40.0",
]
readme = "README.md"
license = { text = "MIT" }

[project.urls]
Homepage = "https://example.com"

[project.scripts]
panekmodel2 = "panekmodel2.cli:app"

[build-system]
requires = ["setuptools>=67.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
